This is a poster paper submitted into AIED. This paper is about QA system design for university.
In this paper, we propose a chatbot based unified system to tackle two major problems in the university domain. We present a system architecture for faster retrieval of answers to the questions asked by students’ in university domain. Another contribution in this paper is the chatbot based automated real- and periodical-time course evaluation system for helping students’ and teachers’ by improving their course quality